Model Sampling

Defining Models

Defining Samplers

Obtaining Samples

1: Calling Sampler Types

Any object of type SisMcmcSampler or SimMcmcSampler can be called like a function, taking as input an SIS (or SIM) model and outputting an MCMC sample. In this case, the sample is returned as part of a specialsed MCMC output type. This is useful traceplots are desired, since one can simply call plot() on the mcmc output.

2: Using draw_sample()

InteractionNetworkModels.draw_sample — Function

draw_sample(
    mcmc::SisMcmcInsertDeleteEdit{T}, 
    model::SIS{T};
    desired_samples::Int=mcmc.desired_samples, 
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::Vector{Path{T}}=get_init(model, mcmc.init)
    )

3: In-Place Sampling

InteractionNetworkModels.draw_sample! — Function

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SisMcmcInsertDeleteEdit, 
    model::SIS;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIS model model::SIS via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SisMcmcInsertDeleteEdit.

draw_sample!(
    sample_out::InteractionSequenceSample, 
    mcmc::SimMcmcInsertDeleteEdit, 
    model::SIM;
    burn_in::Int=mcmc.burn_in,
    lag::Int=mcmc.lag,
    init::InteractionSequence=get_init(model, mcmc.init)
)

Draw sample in-place from given SIM model model::SIM via MCMC algorithm with edit allocation and interaction insertion/deletion, storing output in sample_out::InteractionSequenceSample.

Accepts keyword arguments to change MCMC output, including burn-in, lag and initial values. If not given, these are set to the default values of the passed MCMC sampler mcmc::SimMcmcInsertDeleteEdit.

Plotting